Update allTasks sorting method.
Tasks are sorted by closest to the current days
start time and if the task is past then it
is automatically behind a task that hasnt past.
Past tasks are sorted the same as regular tasks
at end of array. Allows easier viewing of current
tasks with past showing at the end.